requirements.txt
Django==2.2.7
gunicorn==19.9.0
dj-database-url==0.5.0
psycopg2-binary==2.7.6
whitenoise==4.1
django-taggit==1.1.0
django-s3-storage==0.12.5
django-recaptcha==2.0.5
Pillow==6.2.1
